test-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile test-requirements.in
#
appdirs==1.4.3            # via black
async-generator==1.10     # via -r test-requirements.in, pytest-trio, trio
attrs==19.3.0             # via -r test-requirements.in, black, outcome, pytest, trio
black==19.10b0 ; implementation_name == "cpython"  # via -r test-requirements.in
click==7.0                # via black
coverage==4.5.4           # via pytest-cov
entrypoints==0.3          # via flake8
flake8==3.7.9             # via -r test-requirements.in
idna==2.8                 # via trio
mccabe==0.6.1             # via flake8
more-itertools==8.0.2     # via pytest
mypy-extensions==0.4.3    # via mypy, trio-typing
mypy==0.782 ; implementation_name == "cpython"  # via -r test-requirements.in, trio-typing
outcome==1.0.1            # via trio
packaging==19.2           # via pytest
pathspec==0.6.0           # via black
pluggy==0.13.1            # via pytest
py==1.8.0                 # via pytest
pycodestyle==2.5.0        # via flake8
pyflakes==2.1.1           # via flake8
pyparsing==2.4.5          # via packaging
pytest-cov==2.8.1         # via -r test-requirements.in
pytest-faulthandler==2.0.1  # via -r test-requirements.in
pytest-trio==0.5.2        # via -r test-requirements.in
pytest==5.3.1             # via -r test-requirements.in, pytest-cov, pytest-faulthandler, pytest-trio
regex==2019.12.9          # via black
six==1.13.0               # via packaging
sniffio==1.1.0            # via trio
sortedcontainers==2.1.0   # via trio
toml==0.10.0              # via black
trio-typing==0.5.0        # via -r test-requirements.in
trio==0.17.0              # via -r test-requirements.in, pytest-trio, trio-typing
typed-ast==1.4.0 ; implementation_name == "cpython"  # via -r test-requirements.in, black, mypy
typing-extensions==3.7.4.1  # via mypy, trio-typing
wcwidth==0.1.7            # via pytest
